Tests/LibWeb/Text/expected/link-element-rel-preload-load-event.txt

Got load event
Got error event
